requirements.txt
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
distro==1.9.0
docstring_parser==0.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
langsmith==0.7.14
MarkupSafe==3.0.3
oauthlib==3.3.1
orjson==3.11.7
packaging==26.0
pluggy==1.6.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.11.0
PyNaCl==1.6.2
pytest==9.0.2
python-dotenv==1.2.2
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
sniffio==1.3.1
soupsieve==2.8.3
tweepy==4.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
